+++ gcsfs/__init__.py
@@ -15,16 +15,7 @@ logger = logging.getLogger(__name__)
 from .core import GCSFileSystem
 from .mapping import GCSMap
 
-if os.getenv("ENABLE_GCSFS_DUMMY_IO", "false").lower() in ("true", "1"):
-    try:
-        from .dummy_gcsfs import DummyGcsFileSystem as GCSFileSystem
-
-        logger.info("gcsfs dummy-io features enabled via ENABLE_GCSFS_DUMMY_IO.")
-    except ImportError as e:
-        logger.warning(
-            f"ENABLE_GCSFS_DUMMY_IO is set, but failed to import dummy features: {e}"
-        )
-elif os.getenv("GCSFS_EXPERIMENTAL_ZB_HNS_SUPPORT", "true").lower() in ("true", "1"):
+if os.getenv("GCSFS_EXPERIMENTAL_ZB_HNS_SUPPORT", "true").lower() in ("true", "1"):
     try:
         from .extended_gcsfs import ExtendedGcsFileSystem as GCSFileSystem
